bebig-62 feat community find posts, find detail post

+++ src/main/java/beBig/service/CommunityService.java
@@ -6,9 +6,9 @@ import java.util.List;
 
 public interface CommunityService {
     public List<PostVo> showList();
-    public PostVo showDetail(int postId);
+    public PostVo showDetail(Long postId);
     public void write(PostVo post);
-    public void updateLike(int postId);
+    public void updateLike(Long postId);
     public void update(PostVo post);
-    public void delete(int postId);
+    public void delete(Long postId);
 }
